refactor(server): extract socket auth middleware
- keep websocket auth behavior unchanged while isolating the
  Supabase token verification middleware for follow-up refactors
